backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
openai==1.55.3
httpx==0.27.2
python-dotenv==1.0.0
pydantic
langchain
langchain_openai
scikit-learn
psycopg2-binary
langchain-community
langchain-openai
faiss-cpu
pandas
textblob
asyncio